Fixing some errors

+++ src/main/java/icesi/edu/co/repository/AddressRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import icesi.edu.co.person.Address;
 
 @Repository
-public interface AddressRepo extends CrudRepository<Address, Long>{
+public interface AddressRepo extends CrudRepository<Address, Integer>{
 
 }
